Make top sites scrollable, switch to Flow Layout.
There is a bug, collection view is not scrolled on iPads.
It has occurred before the new top sites behavior.
It only affects users with 18+ bookmarks in favorites, iPad only.
